move create blank meta function
