+++ libs/newsletters-data-client/src/lib/newsletter-storage/NewsletterStorage.ts
@@ -19,13 +19,6 @@ export const IMMUTABLE_PROPERTIES: Readonly<string[]> = [
 	'identityName',
 ];
 
-export const makeBlankMeta = (): MetaData => ({
-	createdTimestamp: 0,
-	createdBy: 'unknown',
-	updatedTimestamp: 0,
-	updatedBy: 'unknown',
-});
-
 export abstract class NewsletterStorage {
 	abstract create(
 		draft: DraftNewsletterData,
